prefactor away the CommitReader
