+++ kernel/src/log_reader/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod checkpoint_manifest;
-pub(crate) mod commit;
